pyproject.toml
[tool.poetry]
name = "yatta-tool"
version = "0.1.0"
description = ""
authors = ["Naitian Zhou <naitian@naitian.org>"]
readme = "README.md"
packages = [
  { include = "yatta" },
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.2"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
click = "^8.1.7"
pydantic-settings = "^2.2.1"
sqlmodel = "^0.0.16"
python-multipart = "^0.0.9"
email-validator = "^2.1.1"
werkzeug = "^3.0.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
tqdm = "^4.66.2"

[tool.poetry.scripts]
yatta = "yatta.cli:cli"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
